solved errors in building
