+++ app/(hideNavbar)/school-dashboard/schedule-class/[id]/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-const page = () => {
+const Page = () => {
   return <div>page</div>;
 };
 
-export default page;
+export default Page;
