+++ app/(showNavbar)/find-tutors/[id]/page.tsx
@@ -1,14 +1,12 @@
-
-import SingleTutor from '@/components/SingleTutor'
-import React from 'react'
+import SingleTutor from "@/components/SingleTutor";
+import React from "react";
 
 const singleTutor = () => {
   return (
-    <div className='bg-stone-100 font-subtext'>
-      
-        <SingleTutor/>
+    <div className="bg-stone-100 font-subtext">
+      <SingleTutor />
     </div>
-  )
-}
+  );
+};
 
-export default singleTutor
+export default singleTutor;
