+++ package.json
@@ -34,6 +34,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-datasheet-grid": "^4.11.4",
     "react-date-range": "^2.0.1",
+    "react-datetime-picker": "^6.0.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18",
     "react-icons": "^5.1.0",
@@ -47,6 +48,7 @@
     "@types/bcryptjs": "^2.4.6",
     "@types/node": "^20",
     "@types/react": "^18",
+    "@types/react-date-range": "^1.4.9",
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.1",
